Namelist class in pace.util
